+++ src/App/Theme.lua
@@ -2,52 +2,56 @@ local plugin = script:FindFirstAncestorWhichIsA('Plugin')
 local Studio = settings().Studio
 
 local Argon = script:FindFirstAncestor('Argon')
-local App = Argon.App
-local Components = App.Components
-local Util = Components.Util
 
 local Fusion = require(Argon.Packages.Fusion)
 
-local isState = require(Util.isState)
-
 local Value = Fusion.Value
 
-local COLOR_MAP = {
-	Brand = {
-		Dark = Color3.fromRGB(120, 100, 220),
-		Light = Color3.fromRGB(120, 100, 220),
-	},
-	Background = {
-		Dark = Color3.fromRGB(45, 45, 45),
-		Light = Color3.fromRGB(250, 250, 250),
-	},
-	Border = {
-		Dark = Color3.fromRGB(85, 85, 85),
-		Light = Color3.fromRGB(170, 170, 170),
-	},
-	Text = {
-		Dark = Color3.fromRGB(250, 250, 250),
-		Light = Color3.fromRGB(20, 20, 20),
-	},
-	TextDimmed = {
-		Dark = Color3.fromRGB(160, 160, 160),
-		Light = Color3.fromRGB(90, 90, 90),
-	},
-	Diff = {
-		Add = Color3.fromRGB(80, 220, 100),
-		Update = Color3.fromRGB(100, 200, 230),
-		Remove = Color3.fromRGB(230, 100, 100),
-	},
-}
+local function getColors(isDark: boolean): { [string]: { Color3 | { [string]: Color3 } } }
+	local border = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.Shadow)
+	local background = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.MainBackground)
+
+	return {
+		-- It's actually 130, 120, 230
+		Brand = Color3.fromRGB(120, 110, 220),
+
+		Background = background,
+		Border = border ~= background and border
+			or Studio.Theme:GetColor(Enum.StudioStyleGuideColor.ScrollBarBackground),
+
+		Text = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.BrightText),
+		TextDimmed = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.SubText),
+		TextBranded = isDark and Studio.Theme:GetColor(Enum.StudioStyleGuideColor.BrightText) or background,
+
+		Diff = {
+			Add = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.DiffTextAdditionBackground),
+			Update = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.DiffLineNumSeparatorBackground),
+			Remove = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.DiffTextDeletionBackground),
+		},
+	}
+end
+
+local function init(): Fusion.Value<Color3>
+	return Value(Color3.new())
+end
 
 local Theme = {
+	-- For autocomplete
 	Colors = {
-		Brand = Value(COLOR_MAP.Brand.Dark),
-		Background = Value(COLOR_MAP.Background.Dark),
-		Border = Value(COLOR_MAP.Border.Dark),
-		Text = Value(COLOR_MAP.Text.Dark),
-		TextDimmed = Value(COLOR_MAP.TextDimmed.Dark),
-		Diff = COLOR_MAP.Diff,
+		Brand = init(),
+
+		Background = init(),
+		Border = init(),
+
+		Text = init(),
+		TextDimmed = init(),
+		TextBranded = init(),
+
+		Diff = {
+			Add = init(),
+			Update = init(),
+			Remove = init(),
+		},
 	},
 
 	Fonts = {
@@ -55,6 +59,7 @@ local Theme = {
 		Bold = Font.fromName('Ubuntu', Enum.FontWeight.Bold),
 		Italic = Font.fromName('Ubuntu', Enum.FontWeight.Regular, Enum.FontStyle.Italic),
 		Mono = Font.fromName('RobotoMono'),
+
 		-- required for TextService:GetTextSize()
 		Enums = {
 			Regular = Enum.Font.Ubuntu,
@@ -64,14 +69,23 @@ local Theme = {
 		},
 	},
 
+	BorderThickness = 1,
 	CornerRadius = 6,
 	ListSpacing = 12,
+	WidgetPadding = 16,
 	Padding = 8,
 
-	BorderThickness = 1,
-	WidgetPadding = 16,
-	CompSizeY = 36,
-	TextSize = 20,
+	TextSize = {
+		Large = 20,
+		Medium = 18,
+		Small = 16,
+	},
+
+	CompSizeY = {
+		Large = 36,
+		Medium = 30,
+		Small = 28,
+	},
 
 	SpringSpeed = 30,
 	SpringDamping = 1.5,
@@ -83,13 +97,16 @@ do
 	local function updateTheme()
 		local _, _, v = Studio.Theme:GetColor(Enum.StudioStyleGuideColor.MainBackground):ToHSV()
 		local isDark = v <= 0.6
-		local asString = isDark and 'Dark' or 'Light'
 
 		Theme.IsDark:set(isDark)
 
-		for key, color in pairs(Theme.Colors) do
-			if isState(color) then
-				color:set(COLOR_MAP[key][asString])
+		for key, color in pairs(getColors(isDark)) do
+			if type(color) ~= 'table' then
+				Theme.Colors[key]:set(color)
+			else
+				for subKey, subColor in pairs(color) do
+					Theme.Colors[key][subKey] = subColor
+				end
 			end
 		end
 	end